// src/types/business/goods/goods/goodsTypes.ts
export interface GoodsAddOrEditForm {
  drugGoodsId: string;
  businessDrugGoodsId?: string;
  goodsCategoryId: string;
  specialGoodsType?: number;
  goodsSn: string;
  dosageFormId: string;
  commonName: string;
  goodsName: string;
  unit: string;
  traceCode: string;
  manufacturerId: string;
  originPlace?: string;
  storeLocationId?: string;
  spec?: string;
  goodsPinyin?: string;
  barcode?: string;
  commonNamePinyin?: string;
  marketingAuthorizationHolder?: string;
  isMedicalInsurance?: number;
  isChineseMedicine?: number;
  goodsClassifyId: string;
  prescriptionFlag?: number;
  approvalNo?: string;
  chargeLevel?: number;
  expireWarnDays?: number;
  latestPurchasePrice?: any;
  drugValidDate?: any;
  fileNo?: string;
  latestSupplierId?: string;
  approvalValidEnd?: any;
  salePackageType?: number;
  bigPackageNum?: number;
  bigPackageUnit?: string;
  midPackageNum?: any;
  midPackageUnit?: string;
  smallPackageNum?: any;
  smallPackageUnit?: string;
  nationalCode?: string;
  provinceCode?: string;
  standardPrice: any;
  supplyPrice: any;
  saleUnit?: string;
  splitFlag?: number;
  splitPrice?: number;
  splitUnit?: string;
  splitRatio?: number;
  taxRate?: any;
  isSpecialPrice?: number;
  suggestRetailPrice?: any;
  priceCoefficient?: any;
  firstSupplierId?: string;
  maxRetailPrice?: any;
  wholesalePrice?: any;
  factoryPrice?: any;
  collectPrice?: any;
  mainCategoryAttr?: string;
  funcAttr?: string;
  adminRouteAttr?: number;
  chineseWesternAttr?: number;
  baseDrugAttr?: number;
  abcAttr?: string;
  businessAttr?: string;
  goodsDetailClassify?: string;
  drugAttr3?: string;
  drugAttr4?: string;
  drugAttr5?: string;
  drugAttr6?: string;
  drugAttr7?: string;
  drugAttr8?: string;
  drugAttr9?: string;
  drugAttr10?: string;
  stimulantAttr?: number;
  englishName?: string;
  chemicalName?: string;
  latinName?: string;
  dosage?: string;
  doseUnit?: string;
  frequency?: string;
  cycleDays?: number;
  cycleTimes?: number;
  effectStartDate?: any;
  effectEndDate?: any;
  extendType?: string;
  usageDosage?: string;
  processingMethod?: string;
  efficacy?: string;
  herbFamilySource?: string;
  herbSpeciesSource?: string;
  medicinalPart?: string;
  propertyDescription?: string;
  mainIngredients?: string;
  mainFunction?: string;
  indications?: string;
  status: number;
  certs: Cert[];
}

// 定义状态类型
export interface GoodsAddOrEditState {
  dialogFormVisible: boolean;
  loading: boolean;
  formRef: any;
  form: GoodsAddOrEditForm;
}